Lobby Rework
Gave up on keeping everything in a single scene and gave the lobby its
own scene. Made an array in the gamemanager to store the connected
playerInfos. made the player name in the lobby change to the names of
connected players. Host holds almost all authority in display of this.
Also allowed directx12 and 11 to be selected by the engine if available
due to issues with dx9 and windows 10. Removed AA option in menu to add
permanent SMAA later.