Game Scene and Organization
Created the scene for the actual game, and added a button the host can
hit in the lobby if two players are selected to switch to it. Created a
netGameManager object that will be used for managing the game state and
functions. Added a background image for the game scene (placeholder like
everything else). And finally some file cleanup and organization.