Game UI Work
More work on the ingame ui. Displays names of currently playing player's
with player 1 on the bottom and 2 on the top, along with the names of
their character. Added in logic when the game is started to check if a
character has a character id of 0 and assigns a random character to
them.